+++ flint/src/Flint/Util.hs
@@ -6,6 +6,9 @@ import Flint.Prelude
 
 import Data.Time.Clock (getCurrentTime, diffUTCTime, NominalDiffTime)
 import qualified Data.UUID as UUID
+import Data.IORef (IORef, newIORef, readIORef, writeIORef)
+import qualified Data.Text.IO as TextIO
+import System.IO.Unsafe (unsafePerformIO)
 
 
 incUUID :: UUID -> UUID
@@ -22,3 +25,15 @@ timeIt action = do
   end <- getCurrentTime
   let dt = diffUTCTime end start -- :: NominalDiffTime
   return (result, dt)
+
+{-# NOINLINE samplingTimingEnabled #-}
+samplingTimingEnabled :: IORef Bool
+samplingTimingEnabled = unsafePerformIO $ newIORef False
+
+enableSamplingTiming :: IO ()
+enableSamplingTiming = writeIORef samplingTimingEnabled True
+
+timingLog :: MonadIO m => Text -> m ()
+timingLog msg = liftIO $ readIORef samplingTimingEnabled >>= \case
+  False -> return ()
+  True -> TextIO.hPutStrLn stderr msg